# openapi/employee-oas.yaml
openapi: 3.1.0
info:
  title: Employees API
  description: APIs to manage employees.
  version: 1.0.0

servers:
  - url: https://techshowcase.dev/
    description: Production env
  - url: http://localhost:8080
    description: local env

tags:
  - name: Employee
    description: Operations to manage employees.

paths:
  /api/v1/employees:
    post:
      tags:
        - Employee
      summary: Create an employee
      description: Create an employee with provided details.
      operationId: createEmployee
      requestBody:
        description: Payload with employee detail
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateEmployeeRequest'
      responses:
        '201':
          description: Employee created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateEmployeeResponse'
        '400':
          description: Invalid request payload
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    get:
      tags:
        - Employee
      summary: Search employees
      description: Search one or more employees
      operationId: searchEmployees
      x-spring-paginated: true
      parameters:
      - in: query
        description: Parameters to search employee(s)
        name: employeeSearchParams
        schema:
          $ref: '#/components/schemas/EmployeeSearchParams'
      responses:
        '200':
          description: Employee found successfully
          content:
            application/json:
              schema:
                type: array
                description: List of employees found
                items:
                 $ref: '#/components/schemas/CreateEmployeeResponse'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

components:
  schemas:
    EmployeeSearchParams:
      type: object
      description: Search parameters that can be used to search employee(s).
      properties:
        id:
          type: long
        employeeCode:
          description: Code of the employee
          type: string
        name:
          description: Name of the employee
          type: string
        email:
          description: Email of the employee
          type: string
        department:
          description: Department of the employee
          type: string
    CreateEmployeeRequest:
      type: object
      description: Employee detail to be created.
      required:
        - name
        - employeeCode
        - department
        - email
        - salary
        - dateOfJoining
      properties:
        name:
          type: string
          description: Name of the employee
        employeeCode:
          type: string
          description: code of the employee
          example: code123
        department:
          type: string
          description: department of the employee
          example: IT
        email:
          type: string
          description: email id of the employee
          example: code123
        salary:
          type: number
          description: salary of the employee
          example: 4000.00
        dateOfJoining:
          type: string
          format: date-time
          description: joining date of the employee
    CreateEmployeeResponse:
      type: object
      description: Employee detail to be created.
      properties:
        id:
          type: long
        name:
          type: string
          description: Name of the employee
        employeeCode:
          type: string
          description: code of the employee
          example: code123
        department:
          type: string
          description: department of the employee
        email:
          type: string
          description: email id of the employee
          example: code123
        salary:
          type: number
          description: salary of the employee
          example: 4000.00
        dateOfJoining:
          type: string
          format: date-time
          description: joining date of the employee
        createdAt:
          type: string
          format: date-time
          description: creation date of the employee
        updatedAt:
          type: string
          format: date-time
          description: date that the employee last updated
    ErrorResponse:
      type: object
      description: exception detail
      properties:
        message:
          type: string
          description: error detail
        errors:
          type: array
          items:
            type: string
        timestamp:
          type: string
          format: date-time